separate transport layer better
