+++ src/admin.js
@@ -2,13 +2,13 @@ const invariant = require('invariant');
 const {debug} = require('./utils');
 
 class Admin {
-  constructor(client) {
-    this._client = client;
+  constructor(transport) {
+    this._transport = transport;
   }
   version() {
-    invariant(this._client.isConnected, 'Client should be connected');
+    invariant(this._transport.isConnected, 'Client should be connected');
     debug('request opentmi version');
-    return this._client
+    return this._transport.transport
       .get('/api/v0/version')
       .then(response => response.data);
   }
