+++ src/index.js
@@ -1,15 +1,17 @@
-const Client = require('./client');
+const Authentication = require('./authentication');
 const Admin = require('./admin');
 const Cluster = require('./cluster');
 const Resource = require('./resources');
 const Result = require('./result');
 const utils = require('./utils');
+const Transport = require('./transports');
 
 module.exports = {
-  Client,
+  Authentication,
   Admin,
   Cluster,
   Resource,
   Result,
-  utils
+  utils,
+  Transport
 };
